Update client/main.lua createZones func
fix: onEnter func in createZones lib.zones.sphere was checking if accessPoint.dropPoint or cache.vehicle was valid for returning when it should return when they are not valid. Checks were changed to every parameter needed.

fix: inside func in createZones lib.zones.sphere was checking if accessPoint.dropPoint or cache.vehicle was valid for returning when it should return when they are not valid. Checks were changed to every parameter needed.

fix: Made sure to check for the existence of accessPoint.coords before using config.drawGarageMarker to not run into any nil issues.